chore(icons): make loader implement isLoading()
