+++ packages/icons/src/PixivIcon.ts
@@ -1,11 +1,9 @@
 import type React from 'react'
 import warning from 'warning'
 import { KnownIconFile } from './filenames'
-import { loaders as defaultLoaders, Loader } from './loaders'
+import { FileLoader, UrlLoader, Loader } from './Loader'
 import { BaseElement, __SERVER__ } from './ssr'
 import { sanitize } from 'dompurify'
-
-const { loadFromFile, loadFromRawUrl } = defaultLoaders
 
 const attributes = ['name', 'scale', 'unsafe-non-guideline-scale'] as const
 
@@ -54,9 +52,7 @@ export class PixivIcon extends BaseElement {
         )
       }
 
-      loaders.set(name, function customLoader() {
-        return loadFromRawUrl(url)
-      })
+      loaders.set(name, new UrlLoader(url))
     })
   }
 
@@ -186,11 +182,15 @@ export class PixivIcon extends BaseElement {
   }
 
   private update() {
-    void this.waitUntilVisible().then(async () => {
-      const { name } = this.props
-      const loader = loaders.get(name) ?? loadFromFile
+    const { name } = this.props
+    const loader = loaders.get(name) ?? new FileLoader(name)
 
-      this.svgContent = await loader(name)
+    if (loader.isLoading()) {
+      return
+    }
+
+    void this.waitUntilVisible().then(async () => {
+      this.svgContent = await loader.call()
       this.render()
     })
   }
